+++ packages/labnative/pages/index.js
@@ -11,6 +11,7 @@ import Rating from './Rating';
 import Tag from './Tag';
 import Input from './Input';
 import Progress from './Progress';
+import TextArea from './TextArea';
 
 export {
   Button,
@@ -26,4 +27,5 @@ export {
   Tag,
   Input,
   Progress,
+  TextArea,
 };