+++ packages/yoga/src/Theme/themes/BaseTheme.js
@@ -1,6 +1,7 @@
 import tokens from '@gympass/yoga-tokens';
 
 import {
+  textarea,
   input,
   progress,
   stepper,
@@ -99,6 +100,14 @@ const BaseTheme = ({ primary, secondary, tertiary }) => {
       fontSizes,
       fontWeights,
     }),
+    textarea: textarea({
+      borders,
+      spacing,
+      colors,
+      radii,
+      fontSizes,
+      fontWeights,
+    }),
   };
 
   return {
@@ -111,6 +120,7 @@ const BaseTheme = ({ primary, secondary, tertiary }) => {
     fonts,
     breakpoints,
     transition,
+    radii,
   };
 };