+++ packages/yoga/src/index.js
@@ -1,4 +1,4 @@
-import ThemeProvider, { themes } from './Theme';
+import ThemeProvider, { FontLoader, themes } from './Theme';
 import Button from './Button';
 import List from './List';
 import Checkbox from './Checkbox';
@@ -12,9 +12,11 @@ import Rating from './Rating';
 import Tag from './Tag';
 import Input from './Input';
 import Progress from './Progress';
+import TextArea from './TextArea';
 
 export {
   ThemeProvider,
+  FontLoader,
   themes,
   Button,
   List,
@@ -34,4 +36,5 @@ export {
   Tag,
   Input,
   Progress,
+  TextArea,
 };