+++ packages/labnative/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.12.0](https://github.com/Gympass/yoga/compare/@gympass/yoga-labnative@0.11.2...@gympass/yoga-labnative@0.12.0) (2020-02-18)
+
+### Features
+
+- **textarea:** create textarea component to native platform ([7c09f12](https://github.com/Gympass/yoga/commit/7c09f121f596acb730ba532c90ac987356f943c2))
+
+## [0.11.2](https://github.com/Gympass/design-system/compare/@gympass/yoga-labnative@0.11.1...@gympass/yoga-labnative@0.11.2) (2020-02-14)
+
+**Note:** Version bump only for package @gympass/yoga-labnative
+
 ## [0.11.1](https://github.com/Gympass/yoga/compare/@gympass/yoga-labnative@0.11.0...@gympass/yoga-labnative@0.11.1) (2020-02-12)
 
 **Note:** Version bump only for package @gympass/yoga-labnative